// MindMapper/mindmapper.client/src/app/api/model/naiveCardDto.ts
/**
 * MindMapper.Server
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface NaiveCardDto { 
    id?: string | null;
    title?: string | null;
    description?: string | null;
    top?: number;
    left?: number;
    canvasStateId?: string | null;
    naiveId?: number;
}